+++ src/app/accounts/components/BeforeLink.tsx
@@ -4,7 +4,7 @@ import { FaChevronRight } from "react-icons/fa";
 const BeforeLinkPage = () => {
   return (
     <div className="p-8 text-white">
-      <Link href="/accounts">
+      <Link href="/banks">
         <div className="w-full h-16 p-6 font-bold bg-box-color rounded-2xl flex items-center justify-between">
           <span>내 계좌 연결하기</span>
           <FaChevronRight className="text-sm"/>
